Added `direct train` CLI interface (Closes #109) (#139)
* Design `direct train` CLI interface
