+++ direct/cli/utils.py
@@ -2,6 +2,7 @@
 # Copyright (c) DIRECT Contributors
 import argparse
 import pathlib
+import sys
 
 from direct.utils.io import check_is_valid_url
 
@@ -13,3 +14,8 @@ def file_or_url(path):
     if path.is_file():
         return path
     raise argparse.ArgumentTypeError(f"{path} is not a valid file or url.")
+
+
+def check_train_val(key, name):
+    if key is not None and len(key) != 2:
+        sys.exit(f"--{name} has to be of the form `train_folder, validation_folder` if a validation folder is set.")
